move areaData to data/
